promanage-backend/requirements.txt
# requirements.txt - Versión moderna con Pydantic 2.x
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dateutil==2.8.2

# Dependencias adicionales para producción (opcional)
# redis==5.0.1
# celery==5.3.4
# sqlalchemy==2.0.23
# alembic==1.13.1